Create nptr_check
Due to the debug functionality in MSVS, when a pointer is deleted, or in other situations, it may not appear as null to the code. It may appear as one of the following:
0xFDFDFDFD	No man's land (normally outside of a process)
0xDDDDDDDD	Freed memory
0xCDCDCDCD	Uninitialized(global)
0xCCCCCCCC	Uninitialized locals(on the stack)

This code will handle the normal null pointer values as well as the ones specific to MSVS debug physical address values.